Fix native picker web loading and lint warnings
